+++ gen_libs.py
@@ -133,6 +133,7 @@ import calendar
 import chardet
 import base64
 import binascii
+import errno
 
 # Local
 try:
@@ -1603,7 +1604,8 @@ def make_dir(dirname):
         status = True
 
     except OSError as err:
-        if err.args[0] == 13 or err.args[0] == 17:
+        if err.errno == errno.EEXIST or err.errno == errno.EACCES:
+#           if err.args[0] == 13 or err.args[0] == 17:
             print("Error:  {0} for {1}".format(err.args[1], dirname))
 
         else: