+++ test/unit/sonarqube_code_coverage.sh
@@ -138,7 +138,6 @@ coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_cond_req.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_cond_req_or.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_default.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_dir_chk.py
-coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_dir_chk_crt.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_dir_crt.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_exist.py
 coverage run -a --source=gen_class test/unit/gen_class/argparser_arg_file_chk.py